Move Zetkin API object interfaces
