--- src/interfaces/ZetkinOrganization.ts
@@ -1,4 +1,0 @@
-export interface ZetkinOrganization {
-    id: number;
-    title: string;
-}
