--- src/interfaces/ZetkinUser.ts
@@ -1,6 +1,0 @@
-export interface ZetkinUser {
-    first_name: string;
-    id: number;
-    last_name: string;
-    username: string;
-}
