+++ src/pages/o/[orgId]/events/[eventId].tsx
@@ -29,8 +29,8 @@ import getOrg from '../../../../fetching/getOrg';
 import { PageWithLayout } from '../../../../types';
 import { scaffold } from '../../../../utils/next';
 import { useEventResponses } from '../../../../hooks';
-import { ZetkinEvent } from '../../../../interfaces/ZetkinEvent';
-import { ZetkinOrganization } from '../../../../interfaces/ZetkinOrganization';
+import { ZetkinEvent } from '../../../../types/zetkin';
+import { ZetkinOrganization } from '../../../../types/zetkin';
 
 const scaffoldOptions = {
     localeScope: [
