+++ src/types/zetkin.ts
@@ -17,3 +17,29 @@ export interface ZetkinEventResponse {
     };
     id: number;
 }
+
+export interface ZetkinEvent {
+    activity: { title: string };
+    campaign: {
+        id: number;
+        title: string;
+    };
+    end_time: string;
+    id: number;
+    info_text: string;
+    location: { title: string };
+    start_time: string;
+    title?: string;
+}
+
+export interface ZetkinUser {
+    first_name: string;
+    id: number;
+    last_name: string;
+    username: string;
+}
+
+export interface ZetkinOrganization {
+    id: number;
+    title: string;
+}
